Split the NN code into two files, training and inference
